feat(angular): convert module-federation-dev-server to executor
