+++ docs/shared/reference/sitemap.md
@@ -323,10 +323,10 @@
       - [ng-packagr-lite](/nx-api/angular/executors/ng-packagr-lite)
       - [package](/nx-api/angular/executors/package)
       - [browser-esbuild](/nx-api/angular/executors/browser-esbuild)
+      - [module-federation-dev-server](/nx-api/angular/executors/module-federation-dev-server)
       - [webpack-browser](/nx-api/angular/executors/webpack-browser)
       - [dev-server](/nx-api/angular/executors/dev-server)
       - [webpack-server](/nx-api/angular/executors/webpack-server)
-      - [module-federation-dev-server](/nx-api/angular/executors/module-federation-dev-server)
       - [module-federation-dev-ssr](/nx-api/angular/executors/module-federation-dev-ssr)
     - [generators](/nx-api/angular/generators)
       - [add-linting](/nx-api/angular/generators/add-linting)
