+++ packages/angular/executors.json
@@ -19,6 +19,11 @@
       "implementation": "./src/executors/browser-esbuild/browser-esbuild.impl",
       "schema": "./src/executors/browser-esbuild/schema.json",
       "description": "Builds your application with esbuild and adds support for incremental builds."
+    },
+    "module-federation-dev-server": {
+      "implementation": "./src/executors/module-federation-dev-server/module-federation-dev-server.impl",
+      "schema": "./src/executors/module-federation-dev-server/schema.json",
+      "description": "The module-federation-dev-server executor is reserved exclusively for use with host Module Federation applications. It allows the user to specify which remote applications should be served with the host."
     }
   },
   "builders": {
@@ -36,11 +41,6 @@
       "implementation": "./src/builders/webpack-server/webpack-server.impl",
       "schema": "./src/builders/webpack-server/schema.json",
       "description": "The `webpack-server` executor is very similar to the standard `server` builder provided by the Angular Devkit. It is usually used in tandem with `@nrwl/angular:webpack-browser` when your Angular application uses a custom webpack configuration and NgUniversal for SSR."
-    },
-    "module-federation-dev-server": {
-      "implementation": "./src/builders/module-federation-dev-server/module-federation-dev-server.impl",
-      "schema": "./src/builders/module-federation-dev-server/schema.json",
-      "description": "The module-federation-dev-server executor is reserved exclusively for use with host Module Federation applications. It allows the user to specify which remote applications should be served with the host."
     },
     "module-federation-dev-ssr": {
       "implementation": "./src/builders/module-federation-dev-ssr/module-federation-dev-ssr.impl",
