+++ packages/angular/src/executors/module-federation-dev-server/lib/index.ts
@@ -1,0 +1,4 @@
+export * from './build-static-remotes';
+export * from './normalize-options';
+export * from './start-dev-remotes';
+export * from './start-static-remotes-file-server';
